create individual files for each function in the database
